+++ itsf_sales_coding_test_6/src/main/product/Chocolate.java
@@ -9,7 +9,7 @@ public class Chocolate extends Product {
 
 	private static final ProductType CHOCOLATE_PRODUCT_TYPE = ProductType.Food;
 	
-	public Chocolate(Integer price, Packaging packaging, Boolean isImported) {
+	public Chocolate(Float price, Packaging packaging, Boolean isImported) {
 		super(price, CHOCOLATE_PRODUCT_TYPE, packaging, isImported);
 	}
